requirements.txt
aniso8601==7.0.0
apns2==0.7.1
appdirs==1.4.4
asgiref==3.2.10
asn1crypto==1.4.0
autopep8==1.5.4
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
colorama==0.4.4
cryptography==3.3.1
distlib==0.3.1
Django==3.1.1
django-bulma==0.6.0
django-extensions==3.0.9
django-push-notifications==2.0.0
django-recaptcha==2.0.6
djangorestframework==3.12.2
environ==1.0
filelock==3.0.12
graphene==2.1.8
graphene-django==2.13.0
graphql-core==2.3.2
graphql-relay==2.0.1
h2==2.6.2
hpack==3.0.0
http-ece==1.1.0
hyper==0.7.0
hyperframe==3.2.0
idna==2.10
Markdown==3.3.3
ordered-set==4.0.2
ply==3.11
promise==2.3
psycopg2==2.8.6
py-vapid==1.7.1
pycairo==1.20.0
pycodestyle==2.6.0
pycparser==2.20
PyJWT==2.0.0
pyOpenSSL==20.0.0
python-decouple==3.3
pytz==2020.1
pywebpush==1.11.0
requests==2.25.1
Rx==1.6.1
singledispatch==3.4.0.3
six==1.15.0
sqlparse==0.3.1
toml==0.10.1
Unidecode==1.1.1
urllib3==1.26.2
virtualenv==20.2.2
